split archive_collectdes out from cache_collectdes
