+++ tokio/connectors/collectd_es.py
@@ -5,10 +5,74 @@ Retrieve data generated by collectd and stored in ElasticSearch
 
 import copy
 import time
+import json
+import tokio
 try:
     from elasticsearch import Elasticsearch
 except ImportError:
     pass
+
+QUERY_DISK_DATA = {
+    "query": {
+        "constant_score": {
+            "filter": {
+                "bool": {
+                    "must": [
+                        {
+                            "range": {
+                                "@timestamp": {}
+                            }
+                        },
+                        {
+                            "prefix": {
+                                "hostname": "bb"
+                            }
+                        },
+                        {
+                            "prefix": {
+                                "plugin_instance": "nvme"
+                            }
+                        },
+                        {
+                            "regexp": {
+                                "collectd_type": "disk_(octets|ops)"
+                            }
+                        },
+                        {
+                            "term": {
+                                "plugin": "disk"
+                            }
+                        }
+                    ]
+                }
+            }
+        }
+    }
+}
+### Only return the following _source fields
+COLLECTD_SOURCE_FILTER = [
+    '@timestamp',
+    'hostname',
+    'plugin',
+    'collectd_type',
+    'type_instance',
+    'plugin_instance',
+    'value',
+    'longterm',
+    'midterm',
+    'shortterm',
+    'majflt',
+    'minflt',
+    'if_octets',
+    'if_packets',
+    'if_errors',
+    'rx',
+    'tx',
+    'read',
+    'write',
+    'io_time',
+]
+
 
 class CollectdEs(object):
     """
@@ -152,6 +216,30 @@ class CollectdEs(object):
         while more_results:
             self.page = self.scroll()
             more_results = process_page(scroll_state)
+
+    def query_disk(self, t_start, t_end, timeout=None):
+        """Query ElasticSearch for collectd disk plugin data
+
+        Args:
+            t_start (datetime): lower bound for query (inclusive)
+            t_end (datetime): upper bound for query (exclusive)
+            timeout (int): timeout for the query (in seconds)
+        """
+        query = build_timeseries_query(QUERY_DISK_DATA, t_start, t_end)
+
+        ### Print query
+        tokio.debug.debug_print(json.dumps(query, indent=4))
+
+        ### Run query
+        time0 = time.time()
+        self.query_and_scroll(
+            query=query,
+            source_filter=COLLECTD_SOURCE_FILTER,
+            filter_function=lambda x: x['hits']['hits'],
+            flush_every=50000,
+            flush_function=lambda x: x,
+        )
+        tokio.debug.debug_print("ElasticSearch query took %s seconds" % (time.time() - time0))
 
 def build_timeseries_query(orig_query, start, end):
     """
